pyproject.toml
[tool.poetry]
name = "oauth-userdb"
version = "1.0.5"
description = ""
authors = ["Jimming Cheng <jimming@gmail.com>"]
readme = "README.md"
packages = [{include = "oauth_userdb"}]

[tool.poetry.dependencies]
python = "^3.9"
requests = ">=2.32.4,<3.0.0"
pyjwt = ">=2.10.1,<3.0.0"
oauthlib = ">=3.3.1,<4.0.0"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pytest = "^8.3.4"
requests-mock = "^1.12.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
